+++ x/bundles/keeper/keeper_suite_stakers_leave_test.go
@@ -219,12 +219,11 @@ var _ = Describe("stakers leave", Ordered, func() {
 
 		totalReward := uint64(s.App().BundlesKeeper.GetStorageCost(s.Ctx()).MulInt64(100).TruncateInt64()) + pool.OperatingCost
 		networkFee := s.App().BundlesKeeper.GetNetworkFee(s.Ctx())
-		commission, _ := sdk.NewDecFromStr(uploader.Commission)
 
 		treasuryReward := uint64(sdk.NewDec(int64(totalReward)).Mul(networkFee).TruncateInt64())
 		totalUploaderReward := totalReward - treasuryReward
 
-		uploaderPayoutReward := uint64(sdk.NewDec(int64(totalUploaderReward)).Mul(commission).TruncateInt64())
+		uploaderPayoutReward := uint64(sdk.NewDec(int64(totalUploaderReward)).Mul(uploader.Commission).TruncateInt64())
 		uploaderDelegationReward := totalUploaderReward - uploaderPayoutReward
 
 		// assert payout transfer