+++ x/stakers/types/message_update_commission.go
@@ -2,39 +2,28 @@ package types
 
 import (
 	"cosmossdk.io/errors"
+	"github.com/KYVENetwork/chain/util"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	errorsTypes "github.com/cosmos/cosmos-sdk/types/errors"
 )
 
-const TypeMsgUpdateCommission = "update_commission"
-
 var _ sdk.Msg = &MsgUpdateCommission{}
 
-func (msg *MsgUpdateCommission) Route() string {
-	return RouterKey
-}
-
-func (msg *MsgUpdateCommission) Type() string {
-	return TypeMsgUpdateCommission
-}
-
+// GetSigners returns the expected signers for a MsgUpdateCommission message.
 func (msg *MsgUpdateCommission) GetSigners() []sdk.AccAddress {
-	creator, err := sdk.AccAddressFromBech32(msg.Creator)
-	if err != nil {
-		panic(err)
-	}
-	return []sdk.AccAddress{creator}
+	validator, _ := sdk.AccAddressFromBech32(msg.Creator)
+	return []sdk.AccAddress{validator}
 }
 
-func (msg *MsgUpdateCommission) GetSignBytes() []byte {
-	bz := ModuleCdc.MustMarshalJSON(msg)
-	return sdk.MustSortJSON(bz)
-}
-
+// ValidateBasic does a sanity check on the provided data.
 func (msg *MsgUpdateCommission) ValidateBasic() error {
-	_, err := sdk.AccAddressFromBech32(msg.Creator)
-	if err != nil {
-		return errors.Wrapf(errorsTypes.ErrInvalidAddress, "invalid creator address (%s)", err)
+	if _, err := sdk.AccAddressFromBech32(msg.Creator); err != nil {
+		return errors.Wrapf(errorsTypes.ErrInvalidAddress, "invalid validator address: %s", err)
 	}
+
+	if util.ValidatePercentage(msg.Commission) != nil {
+		return errors.Wrapf(errorsTypes.ErrInvalidRequest, "invalid commission")
+	}
+
 	return nil
 }