+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 
 - (`x/bundles`) [#19](https://github.com/KYVENetwork/chain/pull/19) Migrate `NetworkFee` param to type `sdk.Dec`.
 - (`x/delegation`) [#19](https://github.com/KYVENetwork/chain/pull/19) Migrate `VoteSlash`, `UploadSlash`, `TimeoutSlash` params to type `sdk.Dec`.
+- (`x/stakers`) [#19](https://github.com/KYVENetwork/chain/pull/19) Migrate protocol validator `Commission` to type `sdk.Dec`.
 
 ## [v1.0.0](https://github.com/KYVENetwork/chain/releases/tag/v1.0.0) - 2023-03-10